demo/llm-browseruse-single-agent/requirements.txt
langchain-google-genai
browser-use
python-dotenv
playwright